[WIP] Making better messagges way for delte user

+++ src/graphql/mutations/user/delete-user.ts
@@ -5,9 +5,21 @@ export const deleteUserMutation = {
     async deleteUser(_, { id }): Promise<typeof user> {
         const repository = getRepository(Entities.user);
         const user = await repository.findOne({ id });
-        await repository.delete({ id });
-        return {
-            ...user,
+
+        const success = 500;
+        const message = 'Error not deleted';
+
+        if (user) {
+            success: 200;
+            message: 'User deleted successfully';
+            await repository.delete({ id });
+        }
+
+        const response = {
+            success: success,
+            deletedUser: user,
         };
+
+        return response;
     },
 };
